Option to add primary menu item instead of submenu Adding better error messages for license activation/deactivation

+++ edd/edd.php
@@ -1,6 +1,6 @@
 <?php
 
-class EDD_ECN_Plugin {
+class EDD_Sample_Plugin {
     private $vars;
 
     function __construct( $vars ) {
@@ -81,14 +81,26 @@ class EDD_ECN_Plugin {
         /**
          * Modify this depending on where you would like your settings page to appear
          */
-        add_submenu_page(
-            $this->get_var( 'parent_menu_slug' ),
-            $this->get_var( 'plugin_title' ),
-            $this->get_var( 'settings_page_title' ),
-            'add_users',
-            $this->get_var( 'admin_page_slug' ),
-            array( $this, 'license_page' )
-        );
+	    if ( $this->get_var( 'parent_menu_slug' ) ) {
+	        add_submenu_page(
+	            $this->get_var( 'parent_menu_slug' ),
+	            $this->get_var( 'plugin_title' ),
+	            $this->get_var( 'settings_page_title' ),
+	            'manage_options',
+	            $this->get_var( 'admin_page_slug' ),
+	            array( $this, 'license_page' )
+	        );
+	    } else {
+		    add_menu_page(
+			    $this->get_var( 'plugin_title' ),
+		        $this->get_var( 'plugin_title' ),
+			    'manage_options',
+		        $this->get_var( 'admin_page_slug' ),
+			    array( $this, 'license_page' ),
+			    null,
+			    41
+		    );
+	    }
     }
 
     function license_page() {
@@ -97,44 +109,45 @@ class EDD_ECN_Plugin {
         ?>
         <div class="wrap">
         <h2><?php echo esc_html( $this->get_var( 'plugin_title' ) ) ?></h2>
+	    <?php settings_errors(); ?>
         <form method="post" action="options.php">
 
             <?php settings_fields( $this->get_var( 'option_group' ) ); ?>
 
-            <p><?php esc_html( sprintf( __( 'Thank you for purchasing %s!  Please enter your license key below.', 'edd-sample' ), $this->get_var( 'plugin_title' ) ) ); ?></p>
+	        <?php if ( 'valid' != get_option( $this->get_var( 'license_status' ) ) ): ?>
+                <p><?php echo esc_html( sprintf( __( 'Thank you for purchasing %s!  Please enter your license key below.', 'edd-sample' ), $this->get_var( 'plugin_title' ) ) ); ?></p>
+            <?php endif; ?>
 
             <table class="form-table">
                 <tbody>
-                <tr valign="top">
-                    <th scope="row" valign="top">
-                        <?php _e( 'License Key' ); ?>
-                    </th>
-                    <td>
-                        <input id="<?= esc_attr( $this->get_var( 'license_key' ) ) ?>" name="<?= $this->get_var( 'license_key' ) ?>" type="text" class="regular-text" value="<?php esc_attr_e( $license ); ?>" />
-                        <label class="description" for="<?= esc_attr( $this->get_var( 'license_key' ) ) ?>"><?php _e('Enter your license key'); ?></label>
-                    </td>
-                </tr>
-                <?php if( false !== $license ) { ?>
-                    <tr valign="top">
-                        <th scope="row" valign="top">
-                            <?php _e( 'Activate License' ); ?>
-                        </th>
-                        <td>
-                            <?php if ( $status !== false && $status == 'valid' ) { ?>
-                                <span style="color:green;"><?php _e( 'active' ); ?></span>
-                                <?php wp_nonce_field( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ); ?>
-                                <input type="submit" class="button-secondary" name="<?= $this->get_var( 'option_group' ) ?>_deactivate" value="<?php _e('Deactivate License'); ?>"/>
-                            <?php } else {
-                                wp_nonce_field( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ); ?>
-                                <input type="submit" class="button-secondary" name="<?= $this->get_var( 'option_group' ) ?>_activate" value="<?php _e('Activate License'); ?>"/>
-                            <?php } ?>
-                        </td>
-                    </tr>
-                <?php } ?>
+	                <tr valign="top">
+	                    <th scope="row" valign="top">
+	                        <?php _e( 'License Key' ); ?>
+	                    </th>
+	                    <td>
+	                        <input id="<?= esc_attr( $this->get_var( 'license_key' ) ) ?>" name="<?= $this->get_var( 'license_key' ) ?>" type="text" class="regular-text" value="<?php esc_attr_e( $license ); ?>" />
+	                        <label class="description" for="<?= esc_attr( $this->get_var( 'license_key' ) ) ?>"><?php _e('Enter your license key'); ?></label>
+	                    </td>
+	                </tr>
+	                <tr valign="top">
+	                    <th scope="row" valign="top">
+	                        <?php _e( 'Activate License' ); ?>
+	                    </th>
+	                    <td>
+	                        <?php if ( $status !== false && $status == 'valid' ) { ?>
+	                            <span style="color:green;"><?php _e( 'active' ); ?></span>
+	                            <?php wp_nonce_field( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ); ?>
+	                            <input type="submit" class="button-secondary" name="<?= $this->get_var( 'option_group' ) ?>_deactivate" value="<?php _e('Deactivate License'); ?>"/>
+	                        <?php } else {
+	                            wp_nonce_field( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ); ?>
+	                            <input type="submit" class="button-secondary" name="<?= $this->get_var( 'option_group' ) ?>_activate" value="<?php _e('Activate License'); ?>"/>
+	                        <?php } ?>
+	                    </td>
+	                </tr>
                 </tbody>
             </table>
 
-            <p><?php sprintf( __( 'Any questions or problems with your license? %sContact us%s!', 'edd-sample' ), '<a href="' . $this->get_var( 'contact_url' ) . '">', '</a>' ); ?></p>
+            <p><?php echo sprintf( esc_html( __( 'Any questions or problems with your license? %sContact us%s!', 'edd-sample' ) ), '<a href="' . $this->get_var( 'contact_url' ) . '">', '</a>' ); ?></p>
         </form>
     <?php
     }
@@ -156,7 +169,7 @@ class EDD_ECN_Plugin {
         // listen for our activate button to be clicked
         if ( isset( $_POST[ $this->get_var( 'option_group' ) . '_activate' ] ) ) {
             // run a quick security check
-            if( ! check_admin_referer( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ) )
+            if ( ! check_admin_referer( $this->get_var( 'option_group' ) . '_nonce', $this->get_var( 'option_group' ) . '_nonce' ) )
                 return; // get out if we didn't click the Activate button
 
             // save the license key to the database
@@ -177,14 +190,27 @@ class EDD_ECN_Plugin {
             $response = wp_remote_post( $this->get_var( 'store_url' ), array( 'timeout' => 15, 'sslverify' => false, 'body' => $api_params ) );
 
             // make sure the response came back okay
-            if ( is_wp_error( $response ) )
-                return false;
+            if ( is_wp_error( $response ) ) {
+	            add_settings_error(
+		            $this->get_var( 'option_group' ),
+		            'activate',
+		            __( 'There was an error activating the license, please verify your license is correct and try again or contact support.', 'edd-sample' )
+	            );
+	            return false;
+            }
 
             // decode the license data
             $license_data = json_decode( wp_remote_retrieve_body( $response ) );
 
             // $license_data->license will be either "valid" or "invalid"
             update_option( $this->get_var( 'license_status' ), $license_data->license );
+	        if ( 'valid' != $license_data->license ) {
+		        add_settings_error(
+			        $this->get_var( 'option_group' ),
+			        'activate',
+			        __( 'There was an error activating the license, please verify your license is correct and try again or contact support.', 'edd-sample' )
+		        );
+	        }
         }
     }
 
@@ -210,13 +236,12 @@ class EDD_ECN_Plugin {
             $response = wp_remote_post( $this->get_var( 'store_url' ), array( 'timeout' => 15, 'sslverify' => false, 'body' => $api_params ) );
 
             // make sure the response came back okay
-            // TODO: Fix this
             if ( is_wp_error( $response ) ) {
-                add_action( 'admin_notices', function() {
-                    echo '<div class="error"><p>' .
-                         __( 'There was an error deactivating the license, please try again or contact support.', 'edd-sample' ) .
-                         '</p></div>';
-                } );
+	            add_settings_error(
+		            $this->get_var( 'option_group' ),
+		            'deactivate',
+		            __( 'There was an error deactivating the license, please try again or contact support.', 'edd-sample' )
+                );
                 return false;
             }
 
@@ -224,8 +249,20 @@ class EDD_ECN_Plugin {
             $license_data = json_decode( wp_remote_retrieve_body( $response ) );
 
             // $license_data->license will be either "deactivated" or "failed"
-	        // the API returns failed if license already expired, so removing the license_status regardless
-            delete_option( $this->get_var( 'license_status' ) );
+	        if ( 'deactivated' == $license_data->license ) {
+		        add_settings_error(
+			        $this->get_var( 'option_group' ),
+			        'deactivate',
+			        __( 'License deactivated', 'edd-sample' )
+		        );
+		        delete_option( $this->get_var( 'license_status' ) );
+	        } else {
+		        add_settings_error(
+			        $this->get_var( 'option_group' ),
+			        'deactivate',
+			        __( 'Unable to deactivate license, please try again or contact support.', 'edd-sample' )
+		        );
+	        }
         }
     }
